+++ queries/models/gtfs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog - gtfs
 
+## [1.4.1] - 2026-07-22
+
+### Adicionado
+
+- Adiciona o teste `test_check_trajeto_alternativo` no modelo `trips_gtfs` para validar a correspondência de `(servico, evento)` com `ordem_servico_trajeto_alternativo_sentido`(https://github.com/RJ-SMTR/pipelines_v3/pull/411)
+
 ## [1.4.0] - 2026-07-13
 
 ### Corrigido